Happysocks proxy support
