+++ bzt/bza.py
@@ -10,14 +10,11 @@ import time
 import traceback
 from collections import OrderedDict
 from functools import wraps
-from urllib.parse import urlencode, urlparse
-from typing import List
-from socketio.exceptions import ConnectionError
-import socketio
+from urllib.parse import urlencode
 
 import requests
 
-from bzt import ManualShutdown, TaurusException, TaurusNetworkError, AutomatedShutdown
+from bzt import ManualShutdown, TaurusException, TaurusNetworkError
 from bzt.engine import ScenarioExecutor
 from bzt.jmx import try_convert
 from bzt.resources.version import VERSION
@@ -822,102 +819,3 @@ class BZAProxy(BZAObject):
     def get_json(self):
         response = self._request(self._get_url("/download?format=json"), raw_result=True)
         return response
-
-
-class HappysocksEngineNamespace(socketio.ClientNamespace):
-    """
-    Listens to socket.io events for engine namespace.
-    """
-    NAMESPACE = "/v1/engine"
-    METRICS_EVENT = 'metrics'
-    CONCURRENCY_EVENT = 'concurrency'
-
-    def __init__(self):
-        super().__init__(HappysocksEngineNamespace.NAMESPACE)
-        self._log = logging.getLogger(self.__class__.__name__)
-
-    def on_connect(self):
-        self._log.debug("Connected to happysocks server")
-
-    def on_disconnect(self):
-        self._log.debug("Disconnected from happysocks server")
-
-    def on_connect_error(self, data):
-        details = ""
-        if isinstance(data, dict) and data.get('message'):
-            details = data.get('message')
-        self._log.error(f"Happysocks connection error: {details}")
-
-    def metrics_callback(self, data):
-        if isinstance(data, dict) and data.get('error'):
-            self._log.error(f"Happysocks rejected data: {data.get('error')}")
-
-
-class HappysocksClient:
-    """
-    Client for happysocks service.
-    """
-    HEADER_AUTH_TOKEN = "x-auth-token"
-    HEADER_BZM_SESSION = "x-bzm-session"
-
-    def __init__(self, happysocks_address: str, session_id: str, session_token: str, verbose_logging=False,
-                 verify_ssl=True, request_timeout=10, connect_timeout=7) -> None:
-        super().__init__()
-        self._log = logging.getLogger(self.__class__.__name__)
-        parsed_url = urlparse(happysocks_address)
-        self._session_token = session_token
-        self._session_id = session_id
-        self._happysocks_address = f"{parsed_url.scheme}://{parsed_url.netloc}"
-        self._socketio_path = f"{parsed_url.path.rstrip('/')}/api-ws"
-        self._connect_timeout = connect_timeout
-        # socketio logging is too verbose by default
-        socketio_logger = logging.getLogger("SocketIO") if verbose_logging else False
-        http_session = requests.Session()
-        http_session.verify = verify_ssl
-        self._sio = socketio.Client(http_session=http_session, logger=socketio_logger, engineio_logger=socketio_logger,
-                                    request_timeout=request_timeout)
-        self._engine_namespace = HappysocksEngineNamespace()
-        self._sio.register_namespace(self._engine_namespace)
-
-    def connect(self):
-        headers = {
-            HappysocksClient.HEADER_BZM_SESSION: self._session_id,
-            HappysocksClient.HEADER_AUTH_TOKEN: self._session_token
-        }
-        full_address = f"{self._happysocks_address}{self._socketio_path}"
-        self._log.info(f"Connecting to happysocks server {full_address}")
-        start_time = time.time()
-        try:
-            self._sio.connect(self._happysocks_address, namespaces=[HappysocksEngineNamespace.NAMESPACE],
-                              transports=['websocket'], socketio_path=self._socketio_path, headers=headers,
-                              wait_timeout=self._connect_timeout)
-            end_time = time.time()
-            self._log.info(f"Connected to happysocks server ({round(end_time - start_time, 2)}s)")
-        except ConnectionError as e:
-            end_time = time.time()
-            raise TaurusNetworkError(
-                f"Failed to connect to happysocks server {full_address} ({round(end_time - start_time, 2)}s)") from e
-
-    def connected(self):
-        """
-        Returns True if the client is connected. Also handles situation when the connection is dropped. Client supports
-        auto reconnect.
-        """
-        return self._sio.connected
-
-    def disconnect(self):
-        self._log.info("Disconnecting from happysocks server")
-        # socketio client may be stuck in _handle_reconnect thread
-        # read/write thread runs only while the client is connected
-        if self._sio._reconnect_abort:
-            self._sio._reconnect_abort.set()
-        self._sio.disconnect()
-        self._log.info("Disconnected from happysocks server")
-
-    def send_engine_metrics(self, metrics_batch: List[dict], event: str):
-        self._log.debug(f"Sending {len(metrics_batch)} metric items to happysocks on {event} event")
-        try:
-            self._sio.emit(event, metrics_batch, HappysocksEngineNamespace.NAMESPACE,
-                           callback=self._engine_namespace.metrics_callback)
-        except BaseException as e:
-            raise TaurusNetworkError(f"Failed to send the following items {len(metrics_batch)} on {event} event") from e
